Development

See merge request rakshithxaloori/proeliumx-api!257

+++ clips/tasks.py
@@ -1,10 +1,11 @@
 from django.conf import settings
 from django.core.files.storage import default_storage
 
+
 from shinobi.celery import app as celery_app
 
 
-from clips.utils import VIDEO_MAX_SIZE_IN_BYTES, VIDEO_FILE_ARGS, create_job
+from clips.utils import VIDEO_MAX_SIZE_IN_BYTES, VIDEO_FILE_ARGS, create_job, s3_client
 from shinobi.utils import get_media_file_url, get_media_file_path
 from clips.models import Clip
 from feed.models import Post
@@ -14,96 +15,94 @@ from profiles.models import Following
 from notification.models import Notification
 
 
-def delete_upload_file(file_path):
-    # Verify that it's uploads/
-    if default_storage.exists(file_path):
-        default_storage.delete(file_path)
+AWS_UPLOADS_BUCKET_NAME = settings.AWS_UPLOADS_BUCKET_NAME
+S3_FILE_UPLOAD_PATH_PREFIX = settings.S3_FILE_UPLOAD_PATH_PREFIX
+S3_FILE_CONVERT_PATH_PREFIX = settings.S3_FILE_CONVERT_PATH_PREFIX
+S3_FILE_THUMBNAIL_PATH_PREFIX = settings.S3_FILE_THUMBNAIL_PATH_PREFIX
+
+
+def _delete_upload_file(file_path: str):
+    try:
+        if file_path.startswith(S3_FILE_UPLOAD_PATH_PREFIX):
+            s3_client.delete_object(Bucket=AWS_UPLOADS_BUCKET_NAME, Key=file_path)
+
+    except s3_client.exceptions.NoSuchKey:
+        pass
+
+
+def _get_upload_size(file_path):
+    try:
+        response = s3_client.get_object_attributes(
+            Bucket=AWS_UPLOADS_BUCKET_NAME,
+            Key=file_path,
+            ObjectAttributes=[
+                "ObjectSize",
+            ],
+        )
+        return response["ObjectSize"]
+    except Exception:
+        return 0
 
 
 @celery_app.task(queue="celery")
-def check_upload_after_delay(clip_pk):
-    try:
-        clip = Clip.objects.get(pk=clip_pk)
-    except Clip.DoesNotExist:
-        return
-    if not clip.upload_verified:
-        file_path = get_media_file_path(clip.url)
-        if default_storage.exists(file_path):
-            if default_storage.size(file_path) > VIDEO_MAX_SIZE_IN_BYTES:
-                delete_upload_file(file_path)
-                clip.delete()
-            else:
-                clip.upload_verified = True
-                clip.save(update_fields=["upload_verified"])
-                create_job(
-                    file_path=file_path,
-                    height=clip.height,
-                    width=clip.width,
-                )
-        else:
-            clip.delete()
-
-
-@celery_app.task(queue="celery")
-def check_upload_successful_task(file_path):
-    # Check if the upload is successful in S3
-    if file_path is None:
+def check_upload_task(upload_path):
+    if upload_path is None:
         return
     try:
-        file_url = get_media_file_url(file_path)
-        clip = Clip.objects.get(url=file_url)
+        clip = Clip.objects.get(
+            upload_path=upload_path, upload_verified=False, convert_verified=False
+        )
     except Clip.DoesNotExist:
         return
 
-    if default_storage.exists(file_path):
-        if default_storage.size(file_path) > VIDEO_MAX_SIZE_IN_BYTES:
-            delete_upload_file(file_path)
+    file_size = _get_upload_size(clip.upload_path)
+
+    if file_size > 0:
+        if file_size > VIDEO_MAX_SIZE_IN_BYTES:
             clip.delete()
         else:
             clip.upload_verified = True
             clip.save(update_fields=["upload_verified"])
             create_job(
-                file_path=file_path,
+                file_path=clip.upload_path,
                 height=clip.height,
                 width=clip.width,
             )
+    else:
+        clip.delete()
 
 
 @celery_app.task(queue="celery")
-def check_compressed_successful_task(input_s3_url, jobID, durationInMs, videoDetails):
-    # s3://plx-dev-static/clips/uploads/20secs.mov
-    if input_s3_url is None:
+def check_convert_successful_task(upload_s3_url, jobID, durationInMs, videoDetails):
+    # s3://plx-dev-uploads/clips/uploads/20secs.mov
+    if upload_s3_url is None:
         return
-    bucket_url = "s3://{bucket_name}/".format(
-        bucket_name=settings.AWS_STORAGE_BUCKET_NAME
+    upload_bucket_url = "s3://{bucket_name}/".format(
+        bucket_name=AWS_UPLOADS_BUCKET_NAME
     )
-    upload_file_key = input_s3_url.split(bucket_url)[1]
+    upload_file_key = upload_s3_url.split(upload_bucket_url)[1]
 
-    upload_filename = input_s3_url.split(settings.S3_FILE_UPLOAD_PATH_PREFIX + "/")[
+    upload_filename = upload_s3_url.split(settings.S3_FILE_UPLOAD_PATH_PREFIX + "/")[
         1
     ].split(".")[0]
 
-    compressed_file_key = "{compressed_prefix}/{filename}_4.mp4".format(
-        compressed_prefix=settings.S3_FILE_COMPRESSED_PATH_PREFIX,
+    convert_file_key = "{convert_prefix}/{filename}_4.mp4".format(
+        convert_prefix=S3_FILE_CONVERT_PATH_PREFIX,
         filename=upload_filename,
     )
 
-    if default_storage.exists(compressed_file_key):
+    if default_storage.exists(convert_file_key):
         # Delete the uploaded file
-        delete_upload_file(upload_file_key)
+        _delete_upload_file(upload_file_key)
 
         try:
-            compressed_file_key = "{}/{}_{}.mp4".format(
-                settings.S3_FILE_COMPRESSED_PATH_PREFIX, upload_filename, "{}"
+            convert_file_key = "{}/{}_{}.mp4".format(
+                S3_FILE_CONVERT_PATH_PREFIX, upload_filename, "{}"
             )
-            file_cdn_url = get_media_file_url(compressed_file_key)
-            # Replace hex symbols
-            file_cdn_url = file_cdn_url.replace("%7B", "{")
-            file_cdn_url = file_cdn_url.replace("%7D", "}")
+            file_cdn_url = get_media_file_url(convert_file_key)
 
-            clip = Clip.objects.get(url=get_media_file_url(upload_file_key))
-
-            clip.compressed_verified = True
+            clip = Clip.objects.get(upload_path=upload_file_key)
+            clip.convert_verified = True
             clip.url = file_cdn_url
             clip.duration = durationInMs / 1000
             clip.job_id = jobID
@@ -112,7 +111,7 @@ def check_compressed_successful_task(input_s3_url, jobID, durationInMs, videoDet
                 clip.height = videoDetails["heightInPx"]
                 clip.save(
                     update_fields=[
-                        "compressed_verified",
+                        "convert_verified",
                         "url",
                         "duration",
                         "width",
@@ -123,61 +122,57 @@ def check_compressed_successful_task(input_s3_url, jobID, durationInMs, videoDet
 
             else:
                 clip.save(
-                    update_fields=["compressed_verified", "url", "duration", "job_id"]
+                    update_fields=["convert_verified", "url", "duration", "job_id"]
                 )
 
             clip_post = clip.clip_post
             send_clip_notifications_task.delay(clip_post.pk)
 
         except Clip.DoesNotExist:
-            print("Clip.DoesNotExist", upload_file_key, compressed_file_key)
-            fileargs = VIDEO_FILE_ARGS
-            for filearg in fileargs:
-                vid_file_key = compressed_file_key.format(filearg)
-                if default_storage.exists(vid_file_key):
-                    default_storage.delete(vid_file_key)
+            print("Clip.DoesNotExist", upload_file_key, convert_file_key)
+            delete_clip_files.delay(upload_file_key, convert_file_key, None)
 
 
 @celery_app.task(queue="celery")
-def delete_invalid_duration_clip(input_s3_url):
+def delete_invalid_duration_clip(upload_s3_url):
     # Delete clip
-    bucket_url = "s3://{bucket_name}/".format(
-        bucket_name=settings.AWS_STORAGE_BUCKET_NAME
+    upload_bucket_url = "s3://{bucket_name}/".format(
+        bucket_name=AWS_UPLOADS_BUCKET_NAME
     )
-    upload_file_key = input_s3_url.split(bucket_url)[1]
-
-    upload_filename = input_s3_url.split(settings.S3_FILE_UPLOAD_PATH_PREFIX + "/")[
-        1
-    ].split(".")[0]
-
-    compressed_file_key = "{}/{}_{}.mp4".format(
-        settings.S3_FILE_COMPRESSED_PATH_PREFIX, upload_filename, "{}"
-    )
+    upload_file_key = upload_s3_url.split(upload_bucket_url)[1]
 
     try:
-        clip = Clip.objects.get(url=get_media_file_url(file_path=upload_file_key))
+        clip = Clip.objects.get(upload_path=upload_file_key)
+        delete_clip_files(
+            clip.upload_path,
+            get_media_file_path(clip.url),
+            get_media_file_path(clip.thumbnail),
+        )
         clip.delete()
-        delete_clip_task.delay(get_media_file_url(compressed_file_key))
     except Clip.DoesNotExist:
         pass
 
 
 @celery_app.task(queue="celery")
-def delete_clip_task(url):
-    file_path = get_media_file_path(url)
-    # Replace hex symbols
-    file_path = file_path.replace("%7B", "{")
-    file_path = file_path.replace("%7D", "}")
+def delete_clip_files(upload_path: str, convert_path: str, tb_path: str):
+    # Delete upload file
+    _delete_upload_file(upload_path)
 
-    if default_storage.exists(file_path):
-        # Upload
-        delete_upload_file(file_path)
-    # Upload and compressed
-    fileargs = VIDEO_FILE_ARGS
-    for filearg in fileargs:
-        vid_file_key = file_path.format(filearg)
-        if default_storage.exists(vid_file_key):
-            default_storage.delete(vid_file_key)
+    # Delete thumbnail
+    if tb_path.startswith(S3_FILE_THUMBNAIL_PATH_PREFIX) and default_storage.exists(
+        tb_path
+    ):
+        default_storage.delete(tb_path)
+
+    # Delete convert files
+    if convert_path is not None and convert_path.startswith(
+        S3_FILE_CONVERT_PATH_PREFIX
+    ):
+        fileargs = VIDEO_FILE_ARGS
+        for filearg in fileargs:
+            vid_file_key = convert_path.format(filearg)
+            if default_storage.exists(vid_file_key):
+                default_storage.delete(vid_file_key)
 
 
 @celery_app.task(queue="celery")
